\name{uniConSigV1.3-package}
\alias{uniConSigV1.3-package}
\alias{uniConSigV1.3}
\docType{package}
\title{
\packageTitle{uniConSigV1.3}
}
\description{
\packageDescription{uniConSigV1.3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{uniConSigV1.3}
\packageIndices{uniConSigV1.3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{uniConSigV1.3}

Maintainer: \packageMaintainer{uniConSigV1.3}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
